refactor: change response service layer
